requirement.txt
streamlit
streamlit-audio-recorder
google-generativeai
requests
python-dotenv